차트 만드는 함수 파일로 분리
